question-service/requirements.txt
# Database
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3

# Configuration
python-dotenv==1.0.0
pydantic==2.10.0
pydantic-settings==2.6.1
pyyaml==6.0.2

# LLM Provider SDKs
openai==1.54.0
anthropic==0.39.0
google-generativeai==0.8.3

# Utilities
requests==2.32.3

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
types-PyYAML
